update:gorm升级相关调整

+++ library/app/response.go
@@ -22,7 +22,7 @@ type ResponseData struct {
 
 type ResponseList struct {
 	ResponseData
-	Total int `json:"total"'`
+	Total int64 `json:"total"'`
 }
 
 func (resp *BaseResponse) SetBaseInfo(err error) {
